+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DZ-sk1bO.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-C6Rfc7Bm.js";import{_ as a}from"./mermaid.core-C_6U9Xcl.js";import"./chunk-5VM5RSS4-BBffdhQX.js";import"./chunk-XXDRQBXY-Z5WpGyVt.js";import"./chunk-POPQ4Y6H-B93Y3eTm.js";import"./chunk-F27PBJKO-Ch9BcnsO.js";import"./channel-SwQ6cDWk.js";import"./index-B1ijhMUf.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};